+++ extensions/adb/src/adb-airplane.ts
@@ -1,21 +1,14 @@
-import { LaunchProps, showHUD } from "@raycast/api";
+import { showHUD } from "@raycast/api";
 import { execSync } from "child_process";
 import checkAdbExists from "./utils";
 
-interface AdbAirplaneArguments {
-  toggle: string;
-}
-
-export default async function airplane(props: LaunchProps<{ arguments: AdbAirplaneArguments }>) {
+export default async function airplane() {
   const adbDir = await checkAdbExists();
-  const enable = props.arguments.toggle === "enable" || props.arguments.toggle === "e";
-  let toggleValue;
-  if (enable) {
-    toggleValue = "enable";
-    await showHUD("✈️ Turning on airplane mode");
-  } else {
-    toggleValue = "disable";
+  const enabled = execSync(`${adbDir} shell settings get global airplane_mode_on`).toString().trim() === "true";
+  if (enabled) {
     await showHUD("✈️ Turning off airplane mode");
+  } else {
+    await showHUD("✈️ Turning on airplane mode");
   }
-  execSync(`${adbDir} shell cmd connectivity airplane-mode ${toggleValue}`);
+  execSync(`${adbDir} shell cmd connectivity airplane-mode ${!enabled}`);
 }